social-flow-backend/testing/integration/api-tests/requirements.txt
# Core test libs
pytest>=8.0
requests>=2.31
pytest-asyncio>=0.21
pydantic>=1.10
pytest-xdist>=3.2
pytest-cov>=4.0
python-dotenv>=1.0

# helpful & optional
tenacity>=8.2
jsonschema>=4.17
locust>=2.20
pytest-dependency>=0.5.1
